Summary: Clean Warnings, Sync application Name and title

Reviewers: tcanabrava, patrickelectric, laysrodrigues, leandrosantiago

Reviewed By: tcanabrava, patrickelectric, laysrodrigues

Subscribers: #atelier_atcore

Tags: #atelier_atcore

Differential Revision: https://phabricator.kde.org/D15861

+++ testclient/mainwindow.h
@@ -53,7 +53,7 @@ public slots:
      * @param  number     : index of sensor
      * @param  temp       : temperature
      */
-    void checkTemperature(uint sensorType, uint number, uint temp);
+    void checkTemperature(uint sensorType, uint number, float temp);
 
 private slots:
     //ButtonEvents
